chapter10-image_caption/requirements.txt
git+https://github.com/pytorch/tnt.git@master
fire
torchvision
visdom
jieba
tqdm